Move most of find_yields into a class in modules
